+++ packages/example-remote/src/pages/UserYonYouCorrelation/Edit/index.jsx
@@ -34,14 +34,14 @@ class Detail extends DataTabContainerSupplement {
   tabList = [
     {
       key: 'basicInfo',
-      show: checkHasAuthority(
+      hidden: !checkHasAuthority(
         accessWayCollection.userYonYouCorrelation.get.permission,
       ),
       tab: '基本信息',
     },
     {
       key: 'operateLog/pageList',
-      show: checkHasAuthority(
+      hidden: !checkHasAuthority(
         accessWayCollection.userYonYouCorrelation.pageListOperateLog.permission,
       ),
       tab: '操作日志',