+++ packages/antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 2.11.119
+
+### Patch Changes
+
+- update
+
 ## 2.11.118
 
 ### Patch Changes